User id added

+++ db/car.go
@@ -24,32 +24,32 @@ func PostcarinDB(car Car) error {
 	}
 	return nil
 }
-func GetcarfromDB(id string) ([]*Car, error) {
+func GetcarfromDB(id, userid string) ([]*Car, error) {
 	database := DBConnect()
 	defer database.Close()
 	car := make([]*Car, 0)
-	query := `SELECT car_id, car_name, car_type, capacity,price,gas_type,steering FROM carrental.cars WHERE car_id= $1`
-	err := database.Select(&car, query, id)
+	query := `SELECT car_id,user_id, car_name, car_type, capacity,price,gas_type,steering FROM carrental.cars WHERE car_id= $1 AND user_id = $2`
+	err := database.Select(&car, query, id, userid)
 	if err != nil {
 		return nil, err
 	}
 	return car, nil
 }
-func DeletecarfromDB(id string) error {
+func DeletecarfromDB(id, userid string) error {
 	database := DBConnect()
 	defer database.Close()
-	query := `DELETE FROM carrental.cars WHERE car_id=$1`
-	_, err := database.Exec(query, id)
+	query := `DELETE FROM carrental.cars WHERE car_id=$1 AND user_id = $2`
+	_, err := database.Exec(query, id, userid)
 	if err != nil {
 		return err
 	}
 	fmt.Println("Car Deleted")
 	return nil
 }
-func UpdatecarinDB(id string, cardata Car) error {
+func UpdatecarinDB(id, userid string, cardata Car) error {
 	database := DBConnect()
 	defer database.Close()
-	query := `UPDATE carrental.cars SET car_id=:car_id,car_name=:car_name,car_type=:car_type,capacity=:capacity,price=:price,gas_type=:gas_type, steering=:steering WHERE car_id=$1`
+	query := `UPDATE carrental.cars SET car_id=:car_id,user_id=:user_id, car_name=:car_name,car_type=:car_type,capacity=:capacity,price=:price,gas_type=:gas_type, steering=:steering WHERE car_id=$1 AND user_id=$2`
 	_, err := database.NamedExec(query, cardata)
 	if err != nil {
 		return err
